fix: make the aislop jb gate honest + clear all 218 findings (#33)
## What

CI's aislop gate was a **false green**: it installed roslynator only (never `JetBrains.ReSharper.GlobalTools`) and never built the solution, so the `jb` (ReSharper) engine never ran and every `jb/*` finding was invisible. The gate passed vacuously.

This PR makes the gate honest and clears everything a real run surfaces.

## Changes

- **CI** (`.gitea/workflows/ci.yml`, from the prior WIP commit): install both roslynator and jb (ReSharper CLT), discover VS18 MSBuild via vswhere, and build the full solution so jb has a real compilation graph.
- **`.aislop/config.yml`**: `jbProjects` scopes jb to the managed projects (excludes the C++ native project; cppcheck owns C++); `failBelow` raised to 100.
- **Cleared all 218 findings** a clean run surfaced:
  - ~150 mechanical: redundant usings/qualifiers/casts/default-args (TFM-conditional usings wrapped in `#if`, not deleted), `partial` for CsWinRT1028, `Frame` to `Border` for MAUI ObsoleteElement, async-void command lambdas routed through a logging `RunCommand` bridge so exceptions surface, unused handler params to discards, a genuine always-true condition + if-ladder in TrayIconService, and `MonitorDrift` EDID now surfaced in the drift CLI output (was set but never read).
  - 3 `AccessToDisposedClosure` + 1 serialization DTO setter: per-case `// ReSharper disable once` with justification (correct patterns jb flags conservatively).
  - ~22 MAUI binding/serialization false positives (jb cannot follow RelativeSource / BindableLayout bindings or System.Text.Json member use): excluded by jb rule-id in `lint.csharp.jbExcludeTypes` (binding inspection + the `.Global` unused-member family only; private `.Local` stays active), rationale documented inline.

## Verification

- `aislop ci .`: **score 100/100, 0 findings** across all engines, 115 files (jb genuinely running).
- Solution build: **0 warnings / 0 errors** (VS18 MSBuild).
- Tests: **226 pass** (Claude.Core 186, DisplayManager.Core 40).
- App smoke: launches clean.


---------

Reviewed-on: https://gitea.llamabox.sticktoitive.net/schoen/MegaSchoen/pulls/33
Co-committed-by: claude-code <claude-code@llamabox.internal>

+++ MegaSchoen/Platforms/Windows/Services/Win32Interop.PInvoke.cs
@@ -4,6 +4,7 @@ namespace MegaSchoen.Platforms.Windows.Services;
 
 /// <summary>
 /// P/Invoke declarations for Win32 APIs used by tray icon, hotkeys, and message window.
+/// </summary>
 static partial class Win32Interop
 {
     // Shell32
@@ -157,7 +158,7 @@ static partial class Win32Interop
         // Letters
         if (key.Length == 1 && char.IsLetter(key[0]))
         {
-            return (uint)char.ToUpperInvariant(key[0]);
+            return char.ToUpperInvariant(key[0]);
         }
 
         // Function keys
